refactor: centralize sanitize_for_log, now_iso, and identity utils
