+++ src/fleet_rlm/utils/__init__.py
@@ -5,8 +5,16 @@ Provides small utility functions that don't belong in runtime/integrations.
 
 from __future__ import annotations
 
+from .identity import owner_fingerprint, sanitize_id, session_key
+from .logging import sanitize_for_log
 from .regex import regex_extract
+from .time import now_iso
 
 __all__ = [
+    "now_iso",
+    "owner_fingerprint",
     "regex_extract",
+    "sanitize_for_log",
+    "sanitize_id",
+    "session_key",
 ]
