+++ src/fleet_rlm/utils/time.py
@@ -1,0 +1,13 @@
+"""Time-related utility helpers."""
+
+from __future__ import annotations
+
+from datetime import datetime, timezone
+
+
+def now_iso() -> str:
+    """Return the current UTC time as an ISO 8601 string."""
+    return datetime.now(tz=timezone.utc).isoformat()
+
+
+__all__ = ["now_iso"]
